# api/openapi.yaml
openapi: 3.0.3
info:
  title: assetagent API
  version: 0.1.0
paths:
  /api/health:
    get:
      operationId: getHealth
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /api/transactions:
    get:
      operationId: getTransactions
      summary: List transactions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: from
          in: query
          description: Booking date lower bound (inclusive), YYYY-MM-DD
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: Booking date upper bound (inclusive), YYYY-MM-DD
          schema:
            type: string
            format: date
        - name: account
          in: query
          description: Filter by order account (exact match)
          schema:
            type: string
        - name: counterparty
          in: query
          description: Filter by counterparty name prefix
          schema:
            type: string
        - name: min_amount
          in: query
          description: Minimum amount (inclusive)
          schema:
            type: string
        - name: max_amount
          in: query
          description: Maximum amount (inclusive)
          schema:
            type: string
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [booking_date, amount, counterparty]
            default: booking_date
        - name: order
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: q
          in: query
          description: Search purpose, counterparty, and booking text
          schema:
            type: string
      responses:
        "200":
          description: Paginated transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/chat:
    post:
      operationId: postChat
      summary: Send a chat message and receive a grounded finance answer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Assistant answer with optional tool call trace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/llm/models:
    get:
      operationId: getLLMModels
      summary: List available LLM models for chat
      responses:
        "200":
          description: Model catalog for the chat composer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LLMModelCatalog"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok
    Pagination:
      type: object
      required:
        - limit
        - offset
        - total
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
          format: int64
    Transaction:
      type: object
      required:
        - id
        - order_account
        - booking_date
        - value_date
        - booking_text
        - purpose
        - creditor_id
        - mandate_reference
        - end_to_end_reference
        - collection_reference
        - direct_debit_original_amount
        - chargeback_expense_reimbursement
        - counterparty
        - amount
        - currency
        - info
      properties:
        id:
          type: string
          format: uuid
        order_account:
          type: string
        booking_date:
          type: string
          format: date
        value_date:
          type: string
          format: date
        booking_text:
          type: string
        purpose:
          type: string
        creditor_id:
          type: string
        mandate_reference:
          type: string
        end_to_end_reference:
          type: string
        collection_reference:
          type: string
        direct_debit_original_amount:
          type: string
        chargeback_expense_reimbursement:
          type: string
        counterparty:
          type: string
        counterparty_iban:
          type: string
          nullable: true
        counterparty_bic:
          type: string
          nullable: true
        amount:
          type: string
        currency:
          type: string
        info:
          type: string
    TransactionListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        pagination:
          $ref: "#/components/schemas/Pagination"
    ChatRequest:
      type: object
      required:
        - messages
      properties:
        messages:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ChatMessage"
        provider:
          type: string
          enum: [ollama, openrouter]
          description: Optional LLM provider override
        model:
          type: string
          description: Optional model override for the selected provider
    ChatMessage:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum: [system, user, assistant]
        content:
          type: string
    ChatResponse:
      type: object
      required:
        - answer
        - tool_calls
      properties:
        answer:
          type: string
        tool_calls:
          type: array
          items:
            $ref: "#/components/schemas/ChatToolCall"
    ChatToolCall:
      type: object
      required:
        - name
        - input
        - result
      properties:
        name:
          type: string
        input:
          type: object
          additionalProperties: true
        result:
          type: object
          additionalProperties: true
    LLMModelSelection:
      type: object
      required:
        - provider
        - model
      properties:
        provider:
          type: string
          enum: [ollama, openrouter]
        model:
          type: string
    LLMModelOption:
      type: object
      required:
        - provider
        - model
        - label
      properties:
        provider:
          type: string
          enum: [ollama, openrouter]
        model:
          type: string
        label:
          type: string
        group:
          type: string
          description: Optional UI group, e.g. Cloud or Local
    LLMModelCatalog:
      type: object
      required:
        - default
        - options
      properties:
        default:
          $ref: "#/components/schemas/LLMModelSelection"
        options:
          type: array
          items:
            $ref: "#/components/schemas/LLMModelOption"