# console/src/api/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: `${string}://${string}` | (string & {});
};

export type Error = {
    error: string;
    message: string;
};

export type HealthResponse = {
    status: string;
};

export type Pagination = {
    limit: number;
    offset: number;
    total: number;
};

export type Transaction = {
    id: string;
    order_account: string;
    booking_date: string;
    value_date: string;
    booking_text: string;
    purpose: string;
    creditor_id: string;
    mandate_reference: string;
    end_to_end_reference: string;
    collection_reference: string;
    direct_debit_original_amount: string;
    chargeback_expense_reimbursement: string;
    counterparty: string;
    counterparty_iban?: string | null;
    counterparty_bic?: string | null;
    amount: string;
    currency: string;
    info: string;
};

export type TransactionListResponse = {
    data: Array<Transaction>;
    pagination: Pagination;
};

export type ChatRequest = {
    messages: Array<ChatMessage>;
    /**
     * Optional LLM provider override
     */
    provider?: 'ollama' | 'openrouter';
    /**
     * Optional model override for the selected provider
     */
    model?: string;
};

export type ChatMessage = {
    role: 'system' | 'user' | 'assistant';
    content: string;
};

export type ChatResponse = {
    answer: string;
    tool_calls: Array<ChatToolCall>;
};

export type ChatToolCall = {
    name: string;
    input: {
        [key: string]: unknown;
    };
    result: {
        [key: string]: unknown;
    };
};

export type LlmModelSelection = {
    provider: 'ollama' | 'openrouter';
    model: string;
};

export type LlmModelOption = {
    provider: 'ollama' | 'openrouter';
    model: string;
    label: string;
    /**
     * Optional UI group, e.g. Cloud or Local
     */
    group?: string;
};

export type LlmModelCatalog = {
    default: LlmModelSelection;
    options: Array<LlmModelOption>;
};

export type GetHealthData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/health';
};

export type GetHealthResponses = {
    /**
     * Service is healthy
     */
    200: HealthResponse;
};

export type GetHealthResponse = GetHealthResponses[keyof GetHealthResponses];

export type GetTransactionsData = {
    body?: never;
    path?: never;
    query?: {
        limit?: number;
        offset?: number;
        /**
         * Booking date lower bound (inclusive), YYYY-MM-DD
         */
        from?: string;
        /**
         * Booking date upper bound (inclusive), YYYY-MM-DD
         */
        to?: string;
        /**
         * Filter by order account (exact match)
         */
        account?: string;
        /**
         * Filter by counterparty name prefix
         */
        counterparty?: string;
        /**
         * Minimum amount (inclusive)
         */
        min_amount?: string;
        /**
         * Maximum amount (inclusive)
         */
        max_amount?: string;
        /**
         * Sort field
         */
        sort?: 'booking_date' | 'amount' | 'counterparty';
        /**
         * Sort direction
         */
        order?: 'asc' | 'desc';
        /**
         * Search purpose, counterparty, and booking text
         */
        q?: string;
    };
    url: '/api/transactions';
};

export type GetTransactionsErrors = {
    /**
     * Invalid request
     */
    400: Error;
};

export type GetTransactionsError = GetTransactionsErrors[keyof GetTransactionsErrors];

export type GetTransactionsResponses = {
    /**
     * Paginated transactions
     */
    200: TransactionListResponse;
};

export type GetTransactionsResponse = GetTransactionsResponses[keyof GetTransactionsResponses];

export type PostChatData = {
    body: ChatRequest;
    path?: never;
    query?: never;
    url: '/api/chat';
};

export type PostChatErrors = {
    /**
     * Invalid request
     */
    400: Error;
    /**
     * Internal error
     */
    500: Error;
};

export type PostChatError = PostChatErrors[keyof PostChatErrors];

export type PostChatResponses = {
    /**
     * Assistant answer with optional tool call trace
     */
    200: ChatResponse;
};

export type PostChatResponse = PostChatResponses[keyof PostChatResponses];

export type GetLlmModelsData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/llm/models';
};

export type GetLlmModelsErrors = {
    /**
     * Internal error
     */
    500: Error;
};

export type GetLlmModelsError = GetLlmModelsErrors[keyof GetLlmModelsErrors];

export type GetLlmModelsResponses = {
    /**
     * Model catalog for the chat composer
     */
    200: LlmModelCatalog;
};

export type GetLlmModelsResponse = GetLlmModelsResponses[keyof GetLlmModelsResponses];