# internal/api/gen/api.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.1 DO NOT EDIT.
package gen

import (
	"errors"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for ChatMessageRole.
const (
	Assistant ChatMessageRole = "assistant"
	System    ChatMessageRole = "system"
	User      ChatMessageRole = "user"
)

// Valid indicates whether the value is a known member of the ChatMessageRole enum.
func (e ChatMessageRole) Valid() bool {
	switch e {
	case Assistant:
		return true
	case System:
		return true
	case User:
		return true
	default:
		return false
	}
}

// Defines values for ChatRequestProvider.
const (
	ChatRequestProviderOllama     ChatRequestProvider = "ollama"
	ChatRequestProviderOpenrouter ChatRequestProvider = "openrouter"
)

// Valid indicates whether the value is a known member of the ChatRequestProvider enum.
func (e ChatRequestProvider) Valid() bool {
	switch e {
	case ChatRequestProviderOllama:
		return true
	case ChatRequestProviderOpenrouter:
		return true
	default:
		return false
	}
}

// Defines values for LLMModelOptionProvider.
const (
	LLMModelOptionProviderOllama     LLMModelOptionProvider = "ollama"
	LLMModelOptionProviderOpenrouter LLMModelOptionProvider = "openrouter"
)

// Valid indicates whether the value is a known member of the LLMModelOptionProvider enum.
func (e LLMModelOptionProvider) Valid() bool {
	switch e {
	case LLMModelOptionProviderOllama:
		return true
	case LLMModelOptionProviderOpenrouter:
		return true
	default:
		return false
	}
}

// Defines values for LLMModelSelectionProvider.
const (
	Ollama     LLMModelSelectionProvider = "ollama"
	Openrouter LLMModelSelectionProvider = "openrouter"
)

// Valid indicates whether the value is a known member of the LLMModelSelectionProvider enum.
func (e LLMModelSelectionProvider) Valid() bool {
	switch e {
	case Ollama:
		return true
	case Openrouter:
		return true
	default:
		return false
	}
}

// Defines values for GetTransactionsParamsSort.
const (
	Amount       GetTransactionsParamsSort = "amount"
	BookingDate  GetTransactionsParamsSort = "booking_date"
	Counterparty GetTransactionsParamsSort = "counterparty"
)

// Valid indicates whether the value is a known member of the GetTransactionsParamsSort enum.
func (e GetTransactionsParamsSort) Valid() bool {
	switch e {
	case Amount:
		return true
	case BookingDate:
		return true
	case Counterparty:
		return true
	default:
		return false
	}
}

// Defines values for GetTransactionsParamsOrder.
const (
	Asc  GetTransactionsParamsOrder = "asc"
	Desc GetTransactionsParamsOrder = "desc"
)

// Valid indicates whether the value is a known member of the GetTransactionsParamsOrder enum.
func (e GetTransactionsParamsOrder) Valid() bool {
	switch e {
	case Asc:
		return true
	case Desc:
		return true
	default:
		return false
	}
}

// ChatMessage defines model for ChatMessage.
type ChatMessage struct {
	Content string          `json:"content"`
	Role    ChatMessageRole `json:"role"`
}

// ChatMessageRole defines model for ChatMessage.Role.
type ChatMessageRole string

// ChatRequest defines model for ChatRequest.
type ChatRequest struct {
	Messages []ChatMessage `json:"messages"`

	// Model Optional model override for the selected provider
	Model *string `json:"model,omitempty"`

	// Provider Optional LLM provider override
	Provider *ChatRequestProvider `json:"provider,omitempty"`
}

// ChatRequestProvider Optional LLM provider override
type ChatRequestProvider string

// ChatResponse defines model for ChatResponse.
type ChatResponse struct {
	Answer    string         `json:"answer"`
	ToolCalls []ChatToolCall `json:"tool_calls"`
}

// ChatToolCall defines model for ChatToolCall.
type ChatToolCall struct {
	Input  map[string]interface{} `json:"input"`
	Name   string                 `json:"name"`
	Result map[string]interface{} `json:"result"`
}

// Error defines model for Error.
type Error struct {
	Error   string `json:"error"`
	Message string `json:"message"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Status string `json:"status"`
}

// LLMModelCatalog defines model for LLMModelCatalog.
type LLMModelCatalog struct {
	Default LLMModelSelection `json:"default"`
	Options []LLMModelOption  `json:"options"`
}

// LLMModelOption defines model for LLMModelOption.
type LLMModelOption struct {
	// Group Optional UI group, e.g. Cloud or Local
	Group    *string                `json:"group,omitempty"`
	Label    string                 `json:"label"`
	Model    string                 `json:"model"`
	Provider LLMModelOptionProvider `json:"provider"`
}

// LLMModelOptionProvider defines model for LLMModelOption.Provider.
type LLMModelOptionProvider string

// LLMModelSelection defines model for LLMModelSelection.
type LLMModelSelection struct {
	Model    string                    `json:"model"`
	Provider LLMModelSelectionProvider `json:"provider"`
}

// LLMModelSelectionProvider defines model for LLMModelSelection.Provider.
type LLMModelSelectionProvider string

// Pagination defines model for Pagination.
type Pagination struct {
	Limit  int   `json:"limit"`
	Offset int   `json:"offset"`
	Total  int64 `json:"total"`
}

// Transaction defines model for Transaction.
type Transaction struct {
	Amount                         string             `json:"amount"`
	BookingDate                    openapi_types.Date `json:"booking_date"`
	BookingText                    string             `json:"booking_text"`
	ChargebackExpenseReimbursement string             `json:"chargeback_expense_reimbursement"`
	CollectionReference            string             `json:"collection_reference"`
	Counterparty                   string             `json:"counterparty"`
	CounterpartyBic                *string            `json:"counterparty_bic,omitempty"`
	CounterpartyIban               *string            `json:"counterparty_iban,omitempty"`
	CreditorId                     string             `json:"creditor_id"`
	Currency                       string             `json:"currency"`
	DirectDebitOriginalAmount      string             `json:"direct_debit_original_amount"`
	EndToEndReference              string             `json:"end_to_end_reference"`
	Id                             openapi_types.UUID `json:"id"`
	Info                           string             `json:"info"`
	MandateReference               string             `json:"mandate_reference"`
	OrderAccount                   string             `json:"order_account"`
	Purpose                        string             `json:"purpose"`
	ValueDate                      openapi_types.Date `json:"value_date"`
}

// TransactionListResponse defines model for TransactionListResponse.
type TransactionListResponse struct {
	Data       []Transaction `json:"data"`
	Pagination Pagination    `json:"pagination"`
}

// GetTransactionsParams defines parameters for GetTransactions.
type GetTransactionsParams struct {
	Limit  *int `form:"limit,omitempty" json:"limit,omitempty"`
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`

	// From Booking date lower bound (inclusive), YYYY-MM-DD
	From *openapi_types.Date `form:"from,omitempty" json:"from,omitempty"`

	// To Booking date upper bound (inclusive), YYYY-MM-DD
	To *openapi_types.Date `form:"to,omitempty" json:"to,omitempty"`

	// Account Filter by order account (exact match)
	Account *string `form:"account,omitempty" json:"account,omitempty"`

	// Counterparty Filter by counterparty name prefix
	Counterparty *string `form:"counterparty,omitempty" json:"counterparty,omitempty"`

	// MinAmount Minimum amount (inclusive)
	MinAmount *string `form:"min_amount,omitempty" json:"min_amount,omitempty"`

	// MaxAmount Maximum amount (inclusive)
	MaxAmount *string `form:"max_amount,omitempty" json:"max_amount,omitempty"`

	// Sort Sort field
	Sort *GetTransactionsParamsSort `form:"sort,omitempty" json:"sort,omitempty"`

	// Order Sort direction
	Order *GetTransactionsParamsOrder `form:"order,omitempty" json:"order,omitempty"`

	// Q Search purpose, counterparty, and booking text
	Q *string `form:"q,omitempty" json:"q,omitempty"`
}

// GetTransactionsParamsSort defines parameters for GetTransactions.
type GetTransactionsParamsSort string

// GetTransactionsParamsOrder defines parameters for GetTransactions.
type GetTransactionsParamsOrder string

// PostChatJSONRequestBody defines body for PostChat for application/json ContentType.
type PostChatJSONRequestBody = ChatRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Send a chat message and receive a grounded finance answer
	// (POST /api/chat)
	PostChat(w http.ResponseWriter, r *http.Request)
	// Health check
	// (GET /api/health)
	GetHealth(w http.ResponseWriter, r *http.Request)
	// List available LLM models for chat
	// (GET /api/llm/models)
	GetLLMModels(w http.ResponseWriter, r *http.Request)
	// List transactions
	// (GET /api/transactions)
	GetTransactions(w http.ResponseWriter, r *http.Request, params GetTransactionsParams)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Send a chat message and receive a grounded finance answer
// (POST /api/chat)
func (_ Unimplemented) PostChat(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Health check
// (GET /api/health)
func (_ Unimplemented) GetHealth(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// List available LLM models for chat
// (GET /api/llm/models)
func (_ Unimplemented) GetLLMModels(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// List transactions
// (GET /api/transactions)
func (_ Unimplemented) GetTransactions(w http.ResponseWriter, r *http.Request, params GetTransactionsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// PostChat operation middleware
func (siw *ServerInterfaceWrapper) PostChat(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostChat(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetHealth operation middleware
func (siw *ServerInterfaceWrapper) GetHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetLLMModels operation middleware
func (siw *ServerInterfaceWrapper) GetLLMModels(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetLLMModels(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTransactions operation middleware
func (siw *ServerInterfaceWrapper) GetTransactions(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params GetTransactionsParams

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", r.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "limit"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "offset", r.URL.Query(), &params.Offset, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "offset"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "offset", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "from" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "from", r.URL.Query(), &params.From, runtime.BindQueryParameterOptions{Type: "string", Format: "date"})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "from"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "from", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "to" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "to", r.URL.Query(), &params.To, runtime.BindQueryParameterOptions{Type: "string", Format: "date"})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "to"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "to", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "account" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "account", r.URL.Query(), &params.Account, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "account"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "account", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "counterparty" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "counterparty", r.URL.Query(), &params.Counterparty, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "counterparty"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "counterparty", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "min_amount" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "min_amount", r.URL.Query(), &params.MinAmount, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "min_amount"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "min_amount", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "max_amount" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "max_amount", r.URL.Query(), &params.MaxAmount, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "max_amount"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "max_amount", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "sort" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "sort", r.URL.Query(), &params.Sort, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "sort"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "sort", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "order" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "order", r.URL.Query(), &params.Order, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "order"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "order", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "q" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "q", r.URL.Query(), &params.Q, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "q"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "q", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTransactions(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/chat", wrapper.PostChat)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/health", wrapper.GetHealth)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/llm/models", wrapper.GetLLMModels)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/transactions", wrapper.GetTransactions)
	})

	return r
}